Added Projectiles
Added Projectiles and some enemies to test on, here's some information on how to change different variables:

Offset: Go to the "Player" and expand it to expose the "FirePoint" object. To adjust it, just move it around. (Note: It needs to be to the right of the sprite)

Cooldown: Click on the "FirePoint" object and adjust from there or open up the "WeaponScript" and change the value of "Cooldown". (Note: Don't touch the m_Cooldown value)

Speed and Damage of the bullet: Go to "Prefabs" -> "Projectiles" -> "Bullet" and change from there or go to the "BulletScript" and change variables "Speed" for speed and "Damage" for damage. (Note: If you're changing the speed of the bullet through the script, don't forget to put an "f" after your desired value.

Enemy health: Go to already placed Enemy on the scene, or go to "Prefabs" -> "Characters" -> "Enemy" and change from there, or through the "EnemyScript" script.